Rename intepretUserArgs.js -> parameters.js
This better reflects the general nature of the module, makes it match with
the file that tests it, and resolves ambiguity between the module and the
function of the same name that it exports.
