+++ test/parameters.js
@@ -4,7 +4,7 @@ var path = require( 'path' );
 
 var temp = require( 'temp' ).track();
 
-var interpretUserArgs = require( '../lib/interpretUserArgs' );
+var parameters = require( '../lib/parameters' );
 
 tape( 'interpretUserArgs() correctly handles arguments', function ( test ){
   var testCases = [
@@ -24,7 +24,7 @@ tape( 'interpretUserArgs() correctly handles arguments', function ( test ){
 
   testCases.forEach( function execTestCase( testCase, ind ){
     test.deepEqual(
-      interpretUserArgs.interpretUserArgs( testCase[ 0 ] ), testCase[ 1 ],
+      parameters.interpretUserArgs( testCase[ 0 ] ), testCase[ 1 ],
       util.format( 'Arguments case %d passes.', ind )
     );
   });
@@ -36,7 +36,7 @@ tape( 'interpretUserArgs() correctly handles arguments', function ( test ){
     [ '--deduplicate', 'package.json' ],
   ];
   badArguments.forEach( function execTestCase( testCase, ind ){
-    var errorObj = interpretUserArgs.interpretUserArgs( testCase );
+    var errorObj = parameters.interpretUserArgs( testCase );
     test.ok(
       'exitCode' in errorObj &&  'errMessage' in errorObj,
       'Invalid arguments yield an error object: ' + ind
@@ -49,7 +49,7 @@ tape('interpretUserArgs returns given path as dirPath', function(test) {
   temp.mkdir('tmpdir', function(err, temporary_dir) {
 
     var input = [temporary_dir];
-    var result = interpretUserArgs.interpretUserArgs(input);
+    var result = parameters.interpretUserArgs(input);
 
     test.equal(result.dirPath, temporary_dir, 'path should be equal to specified path');
     test.end();
@@ -67,7 +67,7 @@ tape('interpretUserArgs returns dir from pelias config if no dir specified on co
     };
 
     var input = [];
-    var result = interpretUserArgs.interpretUserArgs(input, peliasConfig);
+    var result = parameters.interpretUserArgs(input, peliasConfig);
 
     test.equal(result.dirPath, temporary_dir, 'path should be equal to path from config');
     test.end();
@@ -89,7 +89,7 @@ tape('getFileList returns fully qualified path names when config has a files lis
 
     var expected = [path.join(temporary_dir, 'filea.csv'), path.join(temporary_dir, 'fileb.csv')];
 
-    var actual = interpretUserArgs.getFileList(peliasConfig, args);
+    var actual = parameters.getFileList(peliasConfig, args);
 
     test.deepEqual(actual, expected, 'file names should be equal');
     test.end();
